requirements.txt
# Core pipeline
pandas==2.2.2
psycopg[binary]==3.2.1
pyyaml==6.0.2
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.4

# Agent (stretch)
langchain==0.3.0
langchain-groq==0.2.0

# Testing
pytest==8.2.2
pytest-cov==5.0.0
